moved queries stage 1 of refactoring
